refactor(schematic): Update the cli's schematic implementation and tests
