+++ schematics/cli-config/files/ignite-ui-cli.json
